login with okta working up to getting called back.
for whatever reason, the LoginCallback from the okta-react package
was buggy, or i wasnt using it correctly.  when i logged in,
the browser would go through a nonterminating loop going back
and forth from the login-callback and back.  i replaced the
LogginCallback component with a dummy component, and it works fine,
in the sense that i get called back with the code in the url query.